requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
pyodbc==5.2.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4